separate logic from form component to calculator
