+++ src/components/Calculator/Calculator.scss
@@ -40,7 +40,7 @@ form {
 .listsContainer {
   display: flex;
   justify-content: center;
-  align-items: center;
+  align-items: flex-start;
   width: 100%;
   min-height: 50%;
 }
