+++ src/components/Form/Form.js
@@ -4,45 +4,20 @@ import uuid from "uuid";
 import { useForm } from "react-hook-form";
 import "./Form.scss";
 
-import { category, addNewItem, handleSnackBar } from "../../utils";
+import { category } from "../../utils";
 import { validationSchema } from "../../utils/validationSchema";
 import YupValidationResolver from "../../utils/YupValidationResolver";
 import Button from "../Button";
 import Budget from "../Budget";
 
-const Form = ({
-  income,
-  setIncome,
-  expense,
-  setExpenses,
-  setIsOpen,
-  sumOfBudget,
-  setSumOfBudget
-}) => {
+const Form = ({ handleAddNewItem, sumOfBudget }) => {
   const resolver = YupValidationResolver(validationSchema);
   const {
     handleSubmit,
-    reset,
     register,
+    reset,
     formState: { errors }
   } = useForm({ resolver });
-
-  const handleAddNewItem = (data) => {
-    const itemType = data.budgetItemType;
-    const amount = parseInt(data.amount, 10);
-    if (itemType === "income") {
-      const newIncome = addNewItem(data);
-      setIncome([...income, newIncome]);
-      setSumOfBudget(sumOfBudget + amount);
-      handleSnackBar(setIsOpen);
-      return reset();
-    }
-    const newExpense = addNewItem(data);
-    setExpenses([...expense, newExpense]);
-    setSumOfBudget(sumOfBudget + -amount);
-    handleSnackBar(setIsOpen);
-    return reset();
-  };
 
   const errorMessages = Object.keys(errors).length ? (
     <ul className="validationList">
@@ -53,9 +28,14 @@ const Form = ({
       ))}
     </ul>
   ) : null;
+
+  const handleAddNewItemControl = (data) => {
+    handleAddNewItem(data);
+    return reset();
+  };
   return (
     <React.Fragment>
-      <form onSubmit={handleSubmit(handleAddNewItem)}>
+      <form onSubmit={handleSubmit(handleAddNewItemControl)}>
         <label className="radioLabel">
           <input type="radio" value="income" {...register("budgetItemType")} />
           Income
@@ -65,7 +45,6 @@ const Form = ({
           Expense
         </label>
         <input className="budgetInput" type="text" placeholder="Name" {...register("name")} />
-
         <input className="budgetInput" type="number" placeholder="Amount" {...register("amount")} />
         <select className="categoryList" name="category" {...register("category")}>
           {category.map((category) => (
@@ -84,12 +63,6 @@ const Form = ({
 export default Form;
 
 Form.propTypes = {
-  income: PropTypes.array,
-  setIncome: PropTypes.func,
-  expense: PropTypes.array,
-  setExpenses: PropTypes.func,
-  isOpen: PropTypes.bool,
-  setIsOpen: PropTypes.func,
-  sumOfBudget: PropTypes.number,
-  setSumOfBudget: PropTypes.func
+  handleAddNewItem: PropTypes.func,
+  sumOfBudget: PropTypes.number
 };
